feat(auth+elastic)!: добавлен поиск с помощью elasticsearch и реализована авторизация через декораторы

+++ app/routers/orders.py
@@ -1,12 +1,12 @@
-# routes/orders.py
 from typing import List
 
-from fastapi import APIRouter, HTTPException, status
+from fastapi import Request, APIRouter, HTTPException, status
 
+from app.middleware.auth import authorize
+from app.schemas.order import OrderCreate, OrderUpdate, OrderResponse
 from libs.database import SessionDep
 from libs.database.models import Order, OrderState
 from libs.database.repositories import OrderRepository
-from app.schemas.order import OrderCreate, OrderUpdate, OrderResponse
 
 router = APIRouter(
     prefix="/orders",
@@ -15,14 +15,16 @@ router = APIRouter(
 
 
 @router.get("/", response_model=List[OrderResponse])
-async def get_all_products(db: SessionDep):
-    """Получить список всех продуктов."""
-    product_repo = OrderRepository(db)
-    products = await product_repo.get_all()
+@authorize()
+async def get_user_orders(request: Request):
+    """Получить список всех заказов пользователя."""
+    product_repo = OrderRepository(request.state.db)
+    products = await product_repo.get_user_orders(request.state.current_user.id)
     return products
 
 
 @router.post("/", response_model=OrderResponse)
+@authorize()
 async def create_order(order_data: OrderCreate, db: SessionDep):
     order_repo = OrderRepository(db)
     new_order = Order(
@@ -35,13 +37,17 @@ async def create_order(order_data: OrderCreate, db: SessionDep):
 
 
 @router.put("/{order_id}", response_model=OrderResponse)
-async def update_order(order_id: int, data: OrderUpdate, db: SessionDep):
-    order_repo = OrderRepository(db)
+@authorize()
+async def update_order(request: Request, order_id: int, data: OrderUpdate):
+    order_repo = OrderRepository(request.state.db)
     order = await order_repo.get(order_id)
     if not order:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Order not found")
 
-    for field, value in data.dict(exclude_unset=True).items():
+    if order.user_id != request.state.current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authorized")
+
+    for field, value in data.model_dump(exclude_unset=True).items():
         setattr(order, field, value)
 
     await order_repo.update(order_id, order)
@@ -49,11 +55,15 @@ async def update_order(order_id: int, data: OrderUpdate, db: SessionDep):
 
 
 @router.delete("/{order_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_order(order_id: int, db: SessionDep):
-    order_repo = OrderRepository(db)
+@authorize()
+async def delete_order(request: Request, order_id: int):
+    order_repo = OrderRepository(request.state.db)
     order = await order_repo.get(order_id)
     if not order:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Order not found")
+
+    if order.user_id != request.state.current_user.id:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authorized")
 
     await order_repo.delete(order)
     return {"message": "Order deleted successfully"}
